fix(pos): Implement Terminal Tap-to-Pay Integration
Resolves #27332

* Implemented safe fallback to mock tokens in `StripeClient`'s `create_terminal_connection_token` and `create_terminal_payment_intent` when `require_api_key()` returns an error. This enables local development and E2E testing without valid credentials.
* Deleted unused and shadowing implementation in `src/server/integrations/stripe/terminal.rs`.
* Fixed `pos.html` to fetch the terminal connection token before the payment intent, and fully utilize `sync_offline` API when processing mock Tap-to-Pay payments.
* Updated backend unit tests to verify the exact behavior of `terminal_api.rs` and `client.rs`'s handling of Stripe Terminal connections.
