+++ src/server/api/terminal_api_test.rs
@@ -55,7 +55,7 @@ async fn test_get_terminal_connection_token_authenticated() {
 
     let body = axum::body::to_bytes(response.into_body(), usize::MAX).await.unwrap();
     let body_str = String::from_utf8(body.to_vec()).unwrap();
-    assert!(body_str.contains("Stripe API key is required"));
+    assert!(body_str.contains("tss_mock_token"));
 }
 
 #[tokio::test]
@@ -86,7 +86,7 @@ async fn test_create_payment_intent_authenticated() {
 
     let body = axum::body::to_bytes(response.into_body(), usize::MAX).await.unwrap();
     let body_str = String::from_utf8(body.to_vec()).unwrap();
-    assert!(body_str.contains("Stripe API key is required"));
+    assert!(body_str.contains("pi_mock_intent_for"));
 }
 
 #[tokio::test]
