--- src/server/integrations/stripe/terminal.rs
@@ -1,54 +1,0 @@
-use crate::integrations::stripe::client::StripeClient;
-
-impl StripeClient {
-    pub async fn create_terminal_connection_token(&self, tenant_id: &str) -> Result<String, String> {
-
-        // In a real implementation, this would make an HTTP POST to Stripe's /v1/terminal/connection_tokens
-        // endpoint. Since we're mocking external APIs, we return a mock token string here.
-        // We simulate the token being tightly scoped to the tenant for multi-tenant isolation.
-        let mock_token = format!("tss_mock_token_for_{}", tenant_id);
-
-        Ok(mock_token)
-    }
-
-    pub async fn create_terminal_payment_intent(
-        &self,
-        tenant_id: &str,
-        amount_cents: i64,
-        currency: &str,
-        _product_id: Option<&str>,
-        _quantity: Option<i32>,
-        _order_id: Option<&str>,
-    ) -> Result<String, String> {
-
-        // In a real implementation, this would make an HTTP POST to Stripe's /v1/payment_intents
-        // endpoint with specific parameters like payment_method_types=["card_present"] and capture_method="manual".
-        // Since we're mocking external APIs, we return a mock intent ID here.
-        let mock_intent = format!("pi_mock_intent_for_{}_{}_{}", tenant_id, amount_cents, currency);
-
-        Ok(mock_intent)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_create_terminal_connection_token() {
-        let client = StripeClient::new("sk_test_123".to_string());
-        let result = client.create_terminal_connection_token("test_tenant").await;
-        assert!(result.is_ok());
-        let token = result.unwrap();
-        assert_eq!(token, "tss_mock_token_for_test_tenant");
-    }
-
-    #[tokio::test]
-    async fn test_create_terminal_payment_intent() {
-        let client = StripeClient::new("sk_test_123".to_string());
-        let result = client.create_terminal_payment_intent("test_tenant", 1500, "usd", None, None, None).await;
-        assert!(result.is_ok());
-        let intent = result.unwrap();
-        assert_eq!(intent, "pi_mock_intent_for_test_tenant_1500_usd");
-    }
-}
